feat: preload project system templates on templates config page
